extract `predicates` into 'bikelines/categories.lua'
